Properly decode null strings (Fixes #18)
When we moved to utilizing the field information in 0.4.0 to correctly
wrap properties using the relevant OMERO rtype an early null check was
inadvertently removed.  This null check was in `omero.rtypes.rtype()`
and partially hidden.  As we are directly using `omero.rtypes.rstring`
now where relevant, and its semantics are to handle `None` as
`RStringI('')`, all `None`'s were being decoded as `RString('')`.

This commit fixes that by adding a null check into `set_property()`.  As
all `None`'s are effectively handled the same and this style mimics the
previous `omero.rtypes.rtype()` semantics the sole test case on the
singular property should suffice.  It is not necessary to test every
string property for round-trip nullability.