Introduce top-level context
To define what the "omero:" field means while keeping as much
backwards compatibility as possible, this introduces a new
argument, `include_context`, to `encode()` which defaults to true.
Previously callers will now receive the new context but internal
uses of `encode()` methods pass `False` meaning that subobjects
will not be littered with additional copies of the context.